Relations associated with a task list established

+++ src/Task_List/Controllers/Task_List_Controller.php
@@ -10,6 +10,7 @@ use League\Fractal\Resource\Collection as Collection;
 use League\Fractal\Pagination\IlluminatePaginatorAdapter;
 use CPM\Transformer_Manager;
 use CPM\Task_List\Transformer\Task_List_Transformer;
+use CPM\Common\Models\Boardable;
 
 class Task_List_Controller {
     use Transformer_Manager;
@@ -77,13 +78,70 @@ class Task_List_Controller {
     }
 
     public function destroy( WP_REST_Request $request ) {
+        // Grab user inputs
         $project_id   = $request->get_param( 'project_id' );
+        $task_list_id = $request->get_param( 'task_list_id' );
+
+        // Select the task list to be deleted
+        $task_list = Task_List::where( 'id', $task_list_id )
+            ->where( 'project_id', $project_id )
+            ->first();
+
+        // Delete relationship with a task list
+        $task_list->boardables()->delete();
+        $comments = $task_list->comments;
+        foreach ( $comments as $comment ) {
+            $comment->replies()->delete();
+            $comment->files()->delete();
+        }
+        $task_list->comments()->delete();
+        $task_list->files()->delete();
+
+        // Delete the task list
+        $task_list->delete();
+    }
+
+    public function attach_users( WP_REST_Request $request ) {
+        $project_id = $request->get_param( 'project_id' );
         $task_list_id = $request->get_param( 'task_list_id' );
 
         $task_list = Task_List::where( 'id', $task_list_id )
             ->where( 'project_id', $project_id )
             ->first();
 
-        $task_list->delete();
+        $user_ids = explode( ',', $request->get_param( 'users' ) );
+
+        if ( !empty( $user_ids ) ) {
+            foreach ( $user_ids as $user_id ) {
+                $data = [
+                    'board_id' => $task_list->id,
+                    'board_type' => 'task-list',
+                    'boardable_id' => $user_id,
+                    'boardable_type' => 'user'
+                ];
+                Boardable::firstOrCreate( $data );
+            }
+        }
+
+        $resource = new Item( $task_list, new Task_List_Transformer );
+
+        return $this->get_response( $resource );
+    }
+
+    public function detach_users( WP_REST_Request $request ) {
+        $project_id = $request->get_param( 'project_id' );
+        $task_list_id = $request->get_param( 'task_list_id' );
+
+        $task_list = Task_List::where( 'id', $task_list_id )
+            ->where( 'project_id', $project_id )
+            ->first();
+
+        $user_ids = explode( ',', $request->get_param( 'users' ) );
+
+        $task_list->users()->whereIn( 'boardable_id', $user_ids )->delete();
+
+        $resource = new Item( $task_list, new Task_List_Transformer );
+
+        return $this->get_response( $resource );
     }
 }
